+++ lib/eventbrite/controllers/auth/authorize/validation.js
@@ -3,4 +3,6 @@ const Joi = require('joi');
 module.exports = () => ({
   dojoId: Joi.string().guid().required(),
   code: Joi.string().required(),
+  userToken: Joi.string().required(),
+  orgId: Joi.string().required(),
 });
