+++ lib/eventbrite/controllers/auth/get/validation.js
@@ -1,9 +1,4 @@
 const Joi = require('joi');
-console.log('VALIDATION');
-
-// module.exports = () => ({
-//   code: Joi.string().required(),
-// });
 
 module.exports = () => ({
   code: Joi.string().required(),
